different return between _write & strlen ?

+++ i2c_descriptor.cpp
@@ -1,5 +1,4 @@
 #include"i2c_descriptor.hpp"
-#include"file_descriptor.hpp"
 
 #include<fcntl.h>//open,write,read
 #include<unistd.h>//close
@@ -20,5 +19,5 @@ i2c_file::i2c_file(const char device[], const int address):file(device)
 
 bool read_byte(const char send[], char buf[], unsigned buf_size)
 {
-  return (_write(buf)==strlen(buf))&&(_read(buf,buf_size)==buf_size);
+  return (_write(buf)==static_cast<int>(strlen(buf)))&&(_read(buf,buf_size)==buf_size);
 }
